create a test_helper module
